* Added limited support for WMC full screen exclusive mode:    - Renderer must be connected to the decoder directly - no intermediate filters.    - Screen is connected to the Intel GPU (decoder shares device with renderer).    - Might only work on a single monitor. * Decoder has exposed its configuration options GetConfig/SetConfig - must be called before initialized. * Padding the image to mod16 width is now off by default. * Decoder can be tested for compatibily with media types via the TestMediaType method. * Fixed leaving IVTC too soon on some cases.
git-svn-id: svn://svn.code.sf.net/p/qsdecoder/code/trunk/IntelQuickSyncDecoder@3 dfccccb7-dd81-40b7-a334-5a7ba89c2b1d

+++ QuickSyncUtils.cpp
@@ -161,7 +161,7 @@ void SetThreadName(LPCSTR szThreadName, DWORD dwThreadID)
     THREADNAME_INFO info;
     info.dwType = 0x1000;
     info.szName = szThreadName;
-    info.dwThreadID = -1;
+    info.dwThreadID = (DWORD)-1;
     info.dwFlags = 0;
 
     __try
